+++ Source/Core/PureES.Tests/Models/TestAggregate.cs
@@ -17,7 +17,7 @@ public partial class TestAggregate
 {
     public required EventEnvelope<Events.Created, Metadata> Created { get; init; }
     public Events.Updated? Updated { get; private set; }
-    public ulong StreamPosition { get; private set; }
+    public uint StreamPosition { get; private set; }
 
     public static Events.Created CreateOn([Command] Commands.Create cmd) => new(cmd.Id, cmd.Value);
 
@@ -52,7 +52,7 @@ public partial class TestAggregate
         return new Events.Updated(current.Created.Event.Id, (int)d);
     }
     
-    public static EventsTransaction CreateTransaction([Command] ulong u) => throw new NotImplementedException();
+    public static EventsTransaction CreateTransaction([Command] uint u) => throw new NotImplementedException();
     
     public EventsTransaction UpdateTransaction([Command] ushort u) => throw new NotImplementedException();
